Parse principle and interest variables to floats and change type of attribute in JSP

+++ src/main/java/com/example/Tomcat_EE_App/Utils.java
@@ -2,7 +2,7 @@ package com.example.Tomcat_EE_App;
 
 public class Utils {
 
-    public static double calculateCompoundInterest(double principleAmount, double interestPercentage, int years, int compoundingPeriod) {
+    public static float calculateCompoundInterest(float principleAmount, float interestPercentage, int years, int compoundingPeriod) {
 
         return Math.round(principleAmount * Math.pow((1 + (interestPercentage / compoundingPeriod)), (compoundingPeriod * years)));
